Bug 1230463: Add a file to quickly set up UDEV rules on Linux r=gerard-majax
Our build instructions tell us to edit this file, but don't tell
us where to get it from! Google's list limits usage to one user.
CyanogenMod's list is embedded in a Wiki. If we have our own file
we can set up everything in a single command for all users.

Signed-off-by: Adam Farden <adam@farden.cz>