Bug 1230736 - Refactor config.sh, purge legacy devices
The device list in config.sh has become wild, unmaintained and unloved.
We have a plethora of legacy devices in this list, when all we care
about right now are emulators, nexuses, flames and Xperias.

Move the legacy devices to legacy-config.sh if anyone really wants to
work on them (Some of these devices can be brought back to life with
a bit of love), and simplify some of the cases that remain in config.sh

Also sort the devices by AOSP base. This will make it relatively clear
for everyone to see exactly how old a device port is, as this generally
correlates with the age of the AOSP base.

Signed-off-by: Adam Farden <adam@farden.cz>