Bug 1211870: Add Sony devices built on CyanogenMod to config menu
Instead of a symlink to a device specific manifest, we use a
symlink to the base-l-cm manifest. The guide will instruct
porters to add their device specific repos to a local manifest.

This will simplify and speed up porting, as there is no need to
fork and create manifests and edit config files which the new
porter is not likely to immediately understand.

The porters can create a device specific manifest after
the port is booting.